add userList.jsp and change file name controller-servlet
